Initial passing test for support for server-side extensions.
